+++ lib/tbank_grpc.rb
@@ -1,8 +1,12 @@
 # frozen_string_literal: true
 
+require 'grpc'
 require 'zeitwerk'
 
+require_relative 'tbank_grpc/errors'
+
 loader = Zeitwerk::Loader.for_gem
+loader.ignore(File.expand_path('tbank_grpc/errors.rb', __dir__))
 loader.setup
 
 module TbankGrpc