+++ lib/tbank_grpc/errors.rb
@@ -1,0 +1,61 @@
+# frozen_string_literal: true
+
+module TbankGrpc
+  class Error < StandardError
+    attr_reader :grpc_code, :context, :tracking_id, :grpc_message, :message
+
+    def self.grpc_code_map
+      @grpc_code_map ||= {
+        GRPC::Core::StatusCodes::INVALID_ARGUMENT => InvalidArgumentError,
+        GRPC::Core::StatusCodes::NOT_FOUND => NotFoundError,
+        GRPC::Core::StatusCodes::ALREADY_EXISTS => AlreadyExistsError,
+        GRPC::Core::StatusCodes::PERMISSION_DENIED => PermissionDeniedError,
+        GRPC::Core::StatusCodes::UNAUTHENTICATED => InvalidTokenError,
+        GRPC::Core::StatusCodes::DEADLINE_EXCEEDED => DeadlineExceededError,
+        GRPC::Core::StatusCodes::UNIMPLEMENTED => UnimplementedError,
+        GRPC::Core::StatusCodes::INTERNAL => InternalError,
+        GRPC::Core::StatusCodes::UNAVAILABLE => UnavailableError
+      }.freeze
+    end
+
+    def initialize(message, context = {})
+      @message = message
+      @grpc_code = context[:grpc_code]
+      @tracking_id = context[:tracking_id]
+      @context = context
+      @grpc_message = context[:message]
+      super(message)
+    end
+
+    def self.from_grpc_error(error)
+      error_class = grpc_code_map[error.code] || Error
+      error_class.new(error.details.to_s, grpc_code: error.code, message: error.details.to_s)
+    end
+
+    def to_s
+      msg = @message
+      msg += " [grpc_message: #{@grpc_message}]" if @grpc_message
+      msg += " [tracking_id: #{@tracking_id}]" if @tracking_id
+      msg += " [code: #{@grpc_code}]" if @grpc_code
+      msg
+    end
+  end
+
+  class ClientError < Error; end
+  class InvalidArgumentError < ClientError; end
+  class NotFoundError < ClientError; end
+  class InvalidTokenError < ClientError; end
+  class PermissionDeniedError < ClientError; end
+  class AlreadyExistsError < ClientError; end
+
+  class ServerError < Error; end
+  class InternalError < ServerError; end
+  class UnavailableError < ServerError; end
+  class DeadlineExceededError < ServerError; end
+  class UnimplementedError < ServerError; end
+
+  class ConnectionError < Error; end
+  class ConnectionFailedError < ConnectionError; end
+  class TLSError < ConnectionError; end
+  class ConfigurationError < Error; end
+end